+++ install-dev/data/xml/authorization_role.xml
@@ -490,9 +490,9 @@
     <authorization_role id="TAB_DEFAULT_READ" slug="ROLE_MOD_TAB_DEFAULT_READ"/>
     <authorization_role id="TAB_DEFAULT_UPDATE" slug="ROLE_MOD_TAB_DEFAULT_UPDATE"/>
     <authorization_role id="TAB_DEFAULT_DELETE" slug="ROLE_MOD_TAB_DEFAULT_DELETE"/>
-    <authorization_role id="TAB_ADMINEXPERIMENTALFEATURE_CREATE" slug="ROLE_MOD_TAB_ADMINEXPERIMENTALFEATURE_CREATE"/>
-    <authorization_role id="TAB_ADMINEXPERIMENTALFEATURE_READ" slug="ROLE_MOD_TAB_ADMINEXPERIMENTALFEATURE_READ"/>
-    <authorization_role id="TAB_ADMINEXPERIMENTALFEATURE_UPDATE" slug="ROLE_MOD_TAB_ADMINEXPERIMENTALFEATURE_UPDATE"/>
-    <authorization_role id="TAB_ADMINEXPERIMENTALFEATURE_DELETE" slug="ROLE_MOD_TAB_ADMINEXPERIMENTALFEATURE_DELETE"/>
+    <authorization_role id="TAB_ADMINFEATUREFLAG_CREATE" slug="ROLE_MOD_TAB_ADMINFEATUREFLAG_CREATE"/>
+    <authorization_role id="TAB_ADMINFEATUREFLAG_READ" slug="ROLE_MOD_TAB_ADMINFEATUREFLAG_READ"/>
+    <authorization_role id="TAB_ADMINFEATUREFLAG_UPDATE" slug="ROLE_MOD_TAB_ADMINFEATUREFLAG_UPDATE"/>
+    <authorization_role id="TAB_ADMINFEATUREFLAG_DELETE" slug="ROLE_MOD_TAB_ADMINFEATUREFLAG_DELETE"/>
   </entities>
 </entity_authorization_role>
