+++ src/PrestaShopBundle/Resources/config/routing/admin/configure/advanced_parameters/_advanced_parameters.yml
@@ -46,6 +46,6 @@ _profiles:
     resource: "profiles.yml"
     prefix: /profiles
 
-_experimental_features:
-    resource: "experimental_features.yml"
-    prefix: /experimental-features
+_feature_flags:
+    resource: "feature_flags.yml"
+    prefix: /feature-flags
