--- src/PrestaShopBundle/Resources/config/routing/admin/configure/advanced_parameters/experimental_features.yml
@@ -1,7 +1,0 @@
-admin_experimental_feature_index:
-    path: /
-    methods: [GET, POST]
-    defaults:
-        _controller: 'PrestaShopBundle:Admin\Configure\AdvancedParameters\ExperimentalFeature:index'
-        _legacy_controller: AdminExperimentalFeature
-        _legacy_link: AdminExperimentalFeature
