+++ src/PrestaShopBundle/Resources/config/routing/admin/configure/advanced_parameters/feature_flags.yml
@@ -1,0 +1,7 @@
+admin_feature_flags_index:
+    path: /
+    methods: [GET, POST]
+    defaults:
+        _controller: 'PrestaShopBundle:Admin\Configure\AdvancedParameters\FeatureFlag:index'
+        _legacy_controller: AdminFeatureFlag
+        _legacy_link: AdminFeatureFlag
