+++ src/PrestaShopBundle/Resources/config/services/bundle/form/form_data_provider.yml
@@ -281,6 +281,6 @@ services:
             - '@prestashop.core.credit_slip.credit_slip_options.configuration'
 
     prestashop.admin.configure.advanced_parameters.feature_flags_form_data_provider:
-        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\ExperimentalFeature\FeatureFlagsFormDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\FeatureFlag\FeatureFlagsFormDataProvider'
         arguments:
             - '@prestashop.core.feature_flags.configuration'
