+++ src/PrestaShopBundle/Resources/config/services/bundle/form/form_handler.yml
@@ -551,5 +551,5 @@ services:
             - '@form.factory'
             - '@prestashop.core.hook.dispatcher'
             - '@prestashop.admin.configure.advanced_parameters.feature_flags_form_data_provider'
-            - 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\ExperimentalFeature\FeatureFlagsType'
+            - 'PrestaShopBundle\Form\Admin\Configure\AdvancedParameters\FeatureFlag\FeatureFlagsType'
             - 'ExperimentalFeature'
